Refactor sp-mvm/tests/assets.rs
A bunch of const variables and obscure enum -> array index conversions
were removed and replaced with Asset and Package structs. All modules
and transactions are available in a separate modules as static
variables. They are not Copy, so passing by reference is required.

Binary data is loaded in runtime using once_cell crate. It allows to
avoid recompilation on changes in assets.